feat(data): chage int to num for every entity and model that has int or double

+++ lib/features/reservation/presentation/screens/reservation_summary_screen.dart
@@ -210,7 +210,7 @@ class _ReservationSummaryScreenState
                             reservationDatetime:
                                 reservation.reservationDatetime,
                             amount: amountAsInt,
-                            numberOfPeople: reservation.numberOfPeople,
+                            numberOfPeople: reservation.numberOfPeople.toInt(),
                           );
                     } else {
                       AppToast.showError(
